[feat(kronos-testing): upgrade Oracle JDBC driver to 23.4.0.24.05]
Update the Oracle JDBC driver version in the kronos-testing module to23.4.0.24.05 to include the latest fixes and improvements. This change
affects the build.gradle.kts file where the testImplementation for the
ojdbc8 dependency is updated.

BREAKING CHANGE: Ensure compatibility with the new JDBC driver version
as this might include changes that could affect existing database
connections or queries.